Move `IMAGE_INFO` into working in debug
An inherent impl on `ObjCObject` seems to make selector registration
happen. This should not be necessary and seems to be from a bug in the
Rust compiler.

I tried making a new project to reproduce the issue being "fixed" here,
but the static seemed to be emitted regardless of what I did to it. So
maybe this is a problem with large crates?